// src/app/_dto/iac3.dto.ts
/*
    Created By  : Arun Joy
    Created On  : 07-01-2020
    Created For : For creating dto for iac3
*/

export class Iac3Dto {

    coCode: string | null;

    ac1Code: string;

    ac2Code: string | null;

    ac3Code: string = "";

    ac3Name: string = "";

    slYn: string = "N";

    bill: string = "N";

    alphabet: string | null = "N";

    job: string = "N";

    acCat: string | null = "N";

}